+++ packages/example-app-base/src/components/Highlight/Highlight.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { registerLanguage, highlightBlock } from 'highlight.js';
 import * as javascript from 'highlight.js/lib/languages/javascript';
 import { BaseComponent } from 'office-ui-fabric-react/lib/Utilities';
-import { createRef } from '@uifabric/utilities/lib/createRef';
+import { createRef } from '@uifabric/utilities';
 
 registerLanguage('javascript', javascript);