+++ packages/experiments/src/components/Chiclet/Chiclet.types.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { ChicletBase } from './Chiclet.base';
-import { IStyleFunction } from '../../Utilities';
+import { IStyleFunctionOrObject } from '../../Utilities';
 import {
   IStyle,
   ITheme
@@ -20,7 +20,7 @@ export interface IChicletProps extends React.Props<ChicletBase> {
   /**
    * Call to provide customized styling that will layer on top of the variant rules.
    */
-  getStyles?: IStyleFunction<IChicletStyleProps, IChicletStyles>;
+  styles?: IStyleFunctionOrObject<IChicletStyleProps, IChicletStyles>;
 
   /**
    * Optional class for chiclet.