+++ packages/experiments/src/components/Nav/SlimNav.tsx
@@ -22,6 +22,9 @@ import { NavLink } from './NavLink';
 const getClassNames = classNamesFunction<INavStyleProps, INavStyles>();
 
 class SlimNavComponent extends NavBase {
+  // store the previous floating nav shown to close when the current floating nav shows up.
+  private _prevFloatingNav: any;
+
   constructor(props: INavProps) {
     super(props);
 
@@ -39,6 +42,10 @@ class SlimNavComponent extends NavBase {
     if (!this.props.groups || this.props.groups.length === 0) {
       return null;
     }
+
+    // reset the flag
+    // on render link, find if there is atleast one hidden link to display "Show more" link
+    this._hasAtleastOneHiddenLink = false;
 
     return (
       <nav role='navigation'>
@@ -95,6 +102,53 @@ class SlimNavComponent extends NavBase {
     ev.stopPropagation();
   }
 
+  private _getFloatingNav(parentElement: HTMLElement | null): HTMLDivElement | undefined {
+    if (!parentElement) {
+      return;
+    }
+
+    return parentElement.querySelector('[data-floating-nav]') as HTMLDivElement;
+  }
+
+  private _onKeyDown(link: INavLink, ev: React.SyntheticEvent<HTMLElement>): void {
+    const nativeEvent = (ev as any);
+    if (nativeEvent.keyCode !== 13) {
+      // accept only enter key to open the floating nav from slim nav
+      return;
+    }
+
+    const a = nativeEvent.target as HTMLElement;
+    const li = a.parentElement;
+    const currentFloatingNav = this._getFloatingNav(li);
+
+    if (!currentFloatingNav) {
+      return;
+    }
+
+    if (this._prevFloatingNav === currentFloatingNav) {
+      // toggle the floating nav
+      if (currentFloatingNav.style && currentFloatingNav.style.display && currentFloatingNav.style.display === 'block') {
+        currentFloatingNav.removeAttribute('style');
+      }
+      else {
+        currentFloatingNav.setAttribute('style', 'display: block');
+      }
+    }
+    else {
+      // prev and current floating navs are different
+      // close the previous if there is one
+      if (this._prevFloatingNav) {
+        this._prevFloatingNav.removeAttribute('style');
+      }
+
+      // open the current one
+      currentFloatingNav.setAttribute('style', 'display: block');
+
+      // store the current as prev
+      this._prevFloatingNav = currentFloatingNav;
+    }
+  }
+
   private _renderCompositeLink(link: INavLink, linkIndex: number, nestingLevel: number): React.ReactElement<{}> | null {
     if (!link) {
       return null;
@@ -108,11 +162,11 @@ class SlimNavComponent extends NavBase {
 
     const isSelected = nestingLevel > 0 && this.isLinkSelected(link, false /* includeChildren */);
     const {
-      getStyles,
+      styles,
       showMore,
       dataHint
     } = this.props;
-    const classNames = getClassNames(getStyles!, { isSelected, nestingLevel });
+    const classNames = getClassNames(styles!, { isSelected, nestingLevel });
     const linkText = this.getLinkText(link, showMore);
 
     return (
@@ -186,11 +240,11 @@ class SlimNavComponent extends NavBase {
     }
 
     const hasChildren = (!!link.links && link.links.length > 0);
-    const { getStyles } = this.props;
-    const classNames = getClassNames(getStyles!, { hasChildren, scrollTop: link.scrollTop });
+    const { styles } = this.props;
+    const classNames = getClassNames(styles!, { hasChildren, scrollTop: link.scrollTop });
 
     return (
-      <div className={ classNames.navFloatingRoot }>
+      <div className={ classNames.navFloatingRoot } data-floating-nav>
         {
           this._renderFloatingLinks([link], 0 /* nestingLevel */)
         }
@@ -206,12 +260,12 @@ class SlimNavComponent extends NavBase {
     const isSelected = this.isLinkSelected(link, true /* includeChildren */);
     const hasChildren = (!!link.links && link.links.length > 0);
     const {
-      getStyles,
+      styles,
       showMore,
       onShowMoreLinkClicked,
       dataHint
     } = this.props;
-    const classNames = getClassNames(getStyles!, { isSelected, hasChildren });
+    const classNames = getClassNames(styles!, { isSelected, hasChildren });
     const linkText = this.getLinkText(link, showMore);
     const onClickHandler = link.isShowMoreLink && onShowMoreLinkClicked ? onShowMoreLinkClicked : this._onLinkClicked.bind(this, link);
 
@@ -221,6 +275,7 @@ class SlimNavComponent extends NavBase {
         key={ link.key || linkIndex }
         onMouseEnter={ this._onLinkMouseEnterOrLeave.bind(this, link) }
         onMouseLeave={ this._onLinkMouseEnterOrLeave.bind(this, link) }
+        onKeyDown={ this._onKeyDown.bind(this, link) }
         title={ linkText }
         className={ classNames.navSlimItemRoot }>
         <NavLink
@@ -229,6 +284,7 @@ class SlimNavComponent extends NavBase {
           target={ link.target }
           dataHint={ dataHint }
           dataValue={ link.key }
+          ariaLabel={ linkText }
           role="menu"
           onClick={ onClickHandler }
           rootClassName={ classNames.navItemRoot }
@@ -257,7 +313,14 @@ class SlimNavComponent extends NavBase {
         {
           links.map((link: INavLink, linkIndex: number) => {
             if (enableCustomization && link.isHidden && !showMore) {
+              // atleast one link is hidden
+              this._hasAtleastOneHiddenLink = true;
+
               // "Show more" overrides isHidden property
+              return null;
+            }
+            else if (link.isShowMoreLink && !this._hasAtleastOneHiddenLink && !showMore) {
+              // there is no hidden link, hide "Show more" link
               return null;
             }
             else {
@@ -275,7 +338,7 @@ class SlimNavComponent extends NavBase {
     }
 
     const {
-      getStyles,
+      styles,
       enableCustomization
     } = this.props;
 
@@ -284,7 +347,7 @@ class SlimNavComponent extends NavBase {
       return null;
     }
 
-    const classNames = getClassNames(getStyles!, {});
+    const classNames = getClassNames(styles!, {});
 
     return (
       <div key={ groupIndex }>