+++ packages/experiments/src/components/Shimmer/ShimmerGap/ShimmerGap.types.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { IStyle, IStyleSet } from '../../../Styling';
-import { IStyleFunction } from '../../../Utilities';
+import { IStyle, IStyleSet, ITheme } from '../../../Styling';
+import { IStyleFunctionOrObject } from '../../../Utilities';
 
 export interface IShimmerGap {
 
@@ -40,9 +40,14 @@ export interface IShimmerGapProps extends React.AllHTMLAttributes<HTMLElement> {
   borderStyle?: IStyleSet;
 
   /**
+  * Theme provided by High-Order Component.
+  */
+  theme?: ITheme;
+
+  /**
    * Call to provide customized styling that will layer on top of the variant rules.
    */
-  getStyles?: IStyleFunction<IShimmerGapStyleProps, IShimmerGapStyles>;
+  styles?: IStyleFunctionOrObject<IShimmerGapStyleProps, IShimmerGapStyles>;
 }
 
 export interface IShimmerGapStyleProps {
@@ -50,6 +55,7 @@ export interface IShimmerGapStyleProps {
   widthInPercentage?: number;
   widthInPixel?: number;
   borderStyle?: IStyleSet;
+  theme: ITheme;
 }
 
 export interface IShimmerGapStyles {