+++ packages/office-ui-fabric-react/src/components/Button/BaseButton.tsx
@@ -9,7 +9,8 @@ import {
   getNativeProps,
   KeyCodes,
   createRef,
-  css
+  css,
+  mergeAriaAttributeValues
 } from '../../Utilities';
 import { Icon } from '../../Icon';
 import { DirectionalHint } from '../../common/DirectionalHint';
@@ -142,16 +143,14 @@ export class BaseButton extends BaseComponent<IBaseButtonProps, IBaseButtonState
       ]);
 
     // Check for ariaDescription, secondaryText or aria-describedby in the native props to determine source of aria-describedby
-    // otherwise default to null.
-    let ariaDescribedBy;
+    // otherwise default to undefined so property does not appear in output.
+    let ariaDescribedBy = undefined;
     if (ariaDescription) {
       ariaDescribedBy = _ariaDescriptionId;
     } else if (secondaryText) {
       ariaDescribedBy = _descriptionId;
     } else if ((nativeProps as any)['aria-describedby']) {
       ariaDescribedBy = (nativeProps as any)['aria-describedby'];
-    } else {
-      ariaDescribedBy = null;
     }
 
     // If an explicit ariaLabel is given, use that as the label and we're done.
@@ -159,12 +158,12 @@ export class BaseButton extends BaseComponent<IBaseButtonProps, IBaseButtonState
     // If any kind of description is given (which will end up as an aria-describedby attribute),
     // set the labelledby element. Otherwise, the button is labeled implicitly by the descendent
     // text on the button (if it exists). Never set both aria-label and aria-labelledby.
-    let ariaLabelledBy = null;
+    let ariaLabelledBy = undefined;
     if (!ariaLabel) {
       if ((nativeProps as any)['aria-labelledby']) {
         ariaLabelledBy = (nativeProps as any)['aria-labelledby'];
       } else if (ariaDescribedBy) {
-        ariaLabelledBy = text ? _labelId : null;
+        ariaLabelledBy = text ? _labelId : undefined;
       }
     }
 
@@ -347,7 +346,7 @@ export class BaseButton extends BaseComponent<IBaseButtonProps, IBaseButtonState
       text = children;
     }
 
-    if (text) {
+    if (this._hasText()) {
       return (
         <div
           key={ this._labelId }
@@ -360,6 +359,13 @@ export class BaseButton extends BaseComponent<IBaseButtonProps, IBaseButtonState
     }
 
     return null;
+  }
+
+  private _hasText(): boolean {
+    // _onRenderTextContents and _onRenderText do not perform the same checks. Below is parity with what _onRenderText used to have
+    // before the refactor that introduced this function. _onRenderTextContents does not require props.text to be undefined in order
+    // for props.children to be used as a fallback. Purely a code maintainability/reuse issue, but logged as Issue #4979
+    return this.props.text !== null && (this.props.text !== undefined || typeof (this.props.children) === 'string');
   }
 
   private _onRenderChildren = (): JSX.Element | null => {
@@ -427,6 +433,13 @@ export class BaseButton extends BaseComponent<IBaseButtonProps, IBaseButtonState
   private _onRenderMenu = (menuProps: IContextualMenuProps): JSX.Element => {
     const { onDismiss = this._dismissMenu } = menuProps;
 
+    // the accessible menu label (accessible name) has a relationship to the button.
+    // If the menu props do not specify an explicit value for aria-label or aria-labelledBy,
+    // AND the button has text, we'll set the menu aria-labelledBy to the text element id.
+    if (!menuProps.ariaLabel && !menuProps.labelElementId && this._hasText()) {
+      menuProps = { ...menuProps, labelElementId: this._labelId };
+    }
+
     return (
       <ContextualMenu
         id={ this._labelId + '-menu' }
@@ -435,7 +448,6 @@ export class BaseButton extends BaseComponent<IBaseButtonProps, IBaseButtonState
         shouldFocusOnContainer={ this.state.menuProps ? this.state.menuProps.shouldFocusOnContainer : undefined }
         className={ css('ms-BaseButton-menuhost', menuProps.className) }
         target={ this._isSplitButton ? this._splitButtonContainer.current : this._buttonElement.current }
-        labelElementId={ this._labelId }
         onDismiss={ onDismiss }
       />
     );
@@ -501,7 +513,7 @@ export class BaseButton extends BaseComponent<IBaseButtonProps, IBaseButtonState
         'data-is-focusable': false
       }
     );
-    const ariaDescribedBy = buttonProps.ariaDescription || '';
+    const ariaDescribedBy = buttonProps.ariaDescription;
 
     if (keytipProps && menuProps) {
       keytipProps = {
@@ -521,7 +533,7 @@ export class BaseButton extends BaseComponent<IBaseButtonProps, IBaseButtonState
             aria-haspopup={ true }
             aria-expanded={ this._isExpanded }
             aria-pressed={ this.props.checked }
-            aria-describedby={ ariaDescribedBy + (keytipAttributes['aria-describedby'] || '') }
+            aria-describedby={ mergeAriaAttributeValues(ariaDescribedBy, keytipAttributes['aria-describedby']) }
             className={ classNames && classNames.splitButtonContainer }
             onKeyDown={ this._onSplitButtonContainerKeyDown }
             onTouchStart={ this._onTouchStart }