+++ packages/office-ui-fabric-react/src/components/Keytip/Keytip.types.ts
@@ -1,6 +1,6 @@
 import { ICalloutProps } from '../../Callout';
 import { IStyle, ITheme } from '../../Styling';
-import { IStyleFunction, IPoint } from '../../Utilities';
+import { IStyleFunctionOrObject, IPoint } from '../../Utilities';
 
 export interface IKeytip {
 }
@@ -83,9 +83,9 @@ export interface IKeytipProps {
   /**
    * Optional styles for the component.
    *
-   * @type {IStyleFunction<IKeytipStyleProps, IKeytipStyles>}
+   * @type {IStyleFunctionOrObject<IKeytipStyleProps, IKeytipStyles>}
    */
-  getStyles?: IStyleFunction<IKeytipStyleProps, IKeytipStyles>;
+  styles?: IStyleFunctionOrObject<IKeytipStyleProps, IKeytipStyles>;
 
   /**
    * Offset x and y for the keytip, added from the top-left corner