+++ packages/office-ui-fabric-react/src/components/KeytipLayer/KeytipLayer.types.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { IStyleFunction } from '../../Utilities';
+import { IStyleFunctionOrObject } from '../../Utilities';
 import { IStyle } from '../../Styling';
 import { IKeytipTransitionKey } from '../../utilities/keytips/IKeytipTransitionKey';
 
@@ -58,9 +58,9 @@ export interface IKeytipLayerProps extends React.Props<IKeytipLayer> {
   onEnterKeytipMode?: () => void;
 
   /**
-   * getStyles function for KeytipLayer
+   * (Optional) Call to provide customized styling.
    */
-  getStyles?: IStyleFunction<IKeytipLayerStyleProps, IKeytipLayerStyles>;
+  styles?: IStyleFunctionOrObject<IKeytipLayerStyleProps, IKeytipLayerStyles>;
 }
 
 export interface IKeytipLayerStyles {